move channel member to channel.go
