--- channel_member.go
@@ -1,17 +1,0 @@
-package stream_chat
-
-import "time"
-
-type ChannelMember struct {
-	UserID      string `json:"user_id,omitempty"`
-	User        *User  `json:"user,omitempty"`
-	IsModerator bool   `json:"is_moderator,omitempty"`
-
-	Invited          bool       `json:"invited,omitempty"`
-	InviteAcceptedAt *time.Time `json:"invite_accepted_at,omitempty"`
-	InviteRejectedAt *time.Time `json:"invite_rejected_at,omitempty"`
-	Role             string     `json:"role,omitempty"`
-
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-}
